backend/task_executor/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# OpenAI
openai==1.3.7

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2

# Logging
structlog==23.2.0


